+++ packages/api/src/StyleOptions.ts
@@ -935,6 +935,18 @@ type StrictStyleOptions = {
    * @default 480
    */
   callURLActionPopupWindowWidth: number | undefined;
+
+  /**
+   * `webchat:callURL` action: origins to open in a trusted popup and have access to `window.opener`, delimited by comma. Same origin is always trusted.
+   *
+   * Default: `undefined`, no cross origins are trusted.
+   *
+   * Security concern: cross origins in this allowlist must be fully trusted, maintained properly, and frequently audited.
+   * Content from these origins could potentially access data and manipulate the page where Web Chat is hosted.
+   *
+   * @default undefined
+   */
+  callURLActionTrustedOrigin: string | undefined;
 };
 
 // StrictStyleOptions is only used internally in Web Chat and for simplifying our code: